Issue 93:	Add created date to label and content
Added created date to label and content class. Based on the way that labels are stored, and that labels contain content, at this point, only the created date on the label will be automatically set when the label is stored. The created date is triggered to be set when toStoreKey is called on the label.